// apps/admin-ui/src/types/Instance.d.ts
interface UbuntuProAccount {
  created_at: string;
  external_account_ids: { IDs: string[]; origin: string }[];
  id: string;
  name: string;
}

interface UbuntuProContract {
  created_at: string;
  id: string;
  name: string;
  products: string[];
  tech_support_level: string;
}

export interface UbuntuProService extends Record<string, unknown> {
  name: string;
  available: string;
  description: string;
  entitled?: string;
  status?: string;
  status_details?: string;
}

interface UbuntuProInfo {
  attached: boolean;
  expires: string | null;
  services: UbuntuProService[];
  techSupportLevel?: string;
  effective?: string;
  contract?: UbuntuProContract;
  account?: UbuntuProAccount;
  result: "success";
}